aws/clean-db/requirements.txt
greenlet==3.0.3
load-dotenv==0.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
SQLAlchemy==2.0.32
typing_extensions==4.12.2
